add skybox

+++ OpenGLCourseApp/SpotLight.h
@@ -6,7 +6,7 @@ class SpotLight :
 public:
 	SpotLight();
 
-	SpotLight(GLfloat shadowWidth, GLfloat shadowHeight,
+	SpotLight(GLuint shadowWidth, GLuint shadowHeight,
 		GLfloat near, GLfloat far, 
 		GLfloat red, GLfloat green, GLfloat blue,
 		GLfloat aIntensity, GLfloat dIntensity,
